Add new plugins and update existing configurations
This commit includes the addition of new plugins such as gitlinker.nvim,
git-dev.nvim, noice.nvim, and virt-column.nvim. It also includes updates
to existing configurations such as lualine, dap, and conform. The
addition of these plugins enhances the functionality of the Neovim
configuration, providing more tools for version control, UI improvements,
and debugging. The updates to the configurations ensure that the plugins
work optimally with the current setup.