feat: Remove virt-column.nvim plugin
The virt-column.nvim plugin was removed from the project. This change
includes the removal of the plugin from the lazy-lock.json file, the
removal of the plugin's configuration from the tui/init.lua file, and
the deletion of the virt-column.lua file. Additionally, the
colorcolumn setting was removed from the settings.lua file.