feat: Add exclude_from_array function in utils
This commit introduces a new function `exclude_from_array` in the
`utils.lua` file. This function takes an array and a set of values to
exclude. It returns a new array that does not include the excluded
values.

In `settings.lua`, the `hide_by_pattern` now uses this new function to
exclude the `.mise.toml` file from the array generated by
`read_gitignore`.